Centralize `{install, revertMocks}()`.
