+++ src/github.com/mozilla-services/pushgo/simplepush/log_emitters_test.go
@@ -7,28 +7,8 @@ package simplepush
 import (
 	"bytes"
 	"encoding/json"
-	"os"
 	"testing"
-	"time"
-
-	"github.com/mozilla-services/pushgo/id"
 )
-
-func installMocks() {
-	osGetPid = func() int { return 1234 }
-	timeNow = func() time.Time { return time.Unix(1257894000, 0).UTC() }
-	idGenerateBytes = func() ([]byte, error) {
-		// d1c7c768-b1be-4c70-93a6-9b52910d4baa.
-		return []byte{0xd1, 0xc7, 0xc7, 0x68, 0xb1, 0xbe, 0x4c, 0x70, 0x93,
-			0xa6, 0x9b, 0x52, 0x91, 0x0d, 0x4b, 0xaa}, nil
-	}
-}
-
-func revertMocks() {
-	osGetPid = os.Getpid
-	timeNow = time.Now
-	idGenerateBytes = id.GenerateBytes
-}
 
 func BenchmarkNewMessage(b *testing.B) {
 	for i := 0; i < b.N; i++ {
@@ -38,7 +18,7 @@ func BenchmarkNewMessage(b *testing.B) {
 		}
 		hm := newHekaMessage()
 		hm.msg.SetID(msgID)
-		hm.msg.SetTimestamp(time.Now().UnixNano())
+		hm.msg.SetTimestamp(timeNow().UnixNano())
 		hm.msg.SetLogger("test-benchmark")
 		hm.msg.SetSeverity(7)
 		hm.msg.SetPayload("Hiya")
